+++ modules/task_2/zaitsev_a_matrix_cannon_block_mult_omp/zaitsev_a_matrix_cannon_block_mult_omp.cpp
@@ -110,7 +110,7 @@ std::vector<std::vector<double>> cannon_mult_alg_Seq(
 std::vector<std::vector<double>> cannon_mult_alg_OMP(
     std::vector<std::vector<double>> first_multiplier,
     std::vector<std::vector<double>> second_multiplier, int thread_nums,
-    size_t size_of_block, size_t count_of_block, size_t size) {
+    size_t size_of_block, int count_of_block, size_t size) {
   std::vector<std::vector<double>> composition(size,
                                                std::vector<double>(size, 0.0));
   int i, j, k, l;
